// src/frontend-applications/agent-portal/modules/object-forms/model/InstructionProperty.ts
/**
 * Copyright (C) 2006-2024 KIX Service Software GmbH, https://www.kixdesk.com
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum InstructionProperty {

    POSSIBLE_VALUES = 'PossibleValues',

    POSSIBLE_VALUES_ADD = 'PossibleValuesAdd',

    POSSIBLE_VALUES_REMOVE = 'PossibleValuesRemove',

    CLEAR = 'Clear',

    SET = 'Set',

    READ_ONLY = 'ReadOnly',

    WRITEABLE = 'Writeable',

    SHOW = 'Show',

    HIDE = 'Hide',

    REQUIRED = 'Required',

    OPTIONAL = 'Optional',

    ENABLE = 'Enable',

    DISABLE = 'Disable',

    MIN_SELECTABLE = 'MinSelectable',

    MAX_SELECTABLE = 'MaxSelectable',

    COUNT_MAX = 'CountMax',

    VALIDATION = 'Validation'

}